requirements.txt
python-telegram-bot>=20.0
# python-dotenv # <-- Comment this line out or delete it
